credit_card_rate/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dateutil==2.8.2

